Reorganize the file and code structure
