+++ lib/src/adapters/dio_adapter.dart
@@ -3,15 +3,22 @@ import 'dart:io';
 import 'package:dio/adapter.dart';
 import 'package:dio/dio.dart';
 import 'package:http_mock_adapter/src/constants.dart';
-import 'package:http_mock_adapter/src/mock_provider.dart';
-import 'package:http_mock_adapter/src/request_handler.dart';
+import 'package:http_mock_adapter/src/handlers/mock_handler.dart';
+import 'package:http_mock_adapter/src/handlers/request_handler.dart';
 import 'package:http_mock_adapter/src/utils.dart';
 
-import 'history.dart';
+import '../history.dart';
 
-class DioAdapter extends HttpClientAdapter with MockProvider, Tracking {
+/// [HttpClientAdapter] extension with data mocking and tracking functionality.
+class DioAdapter extends HttpClientAdapter with Tracking {
+  /// [Dio]`s default HTTP client adapter implementation.
   DefaultHttpClientAdapter _adapter = DefaultHttpClientAdapter();
 
+  /// [MockHandler] abstracts away the functionality necessary to handle
+  /// client requests.
+  final _mockHandler = MockHandler();
+
+  /// Map of <[RequestMethod], <(String) [RequestMethod], [RequestHandler]>>.
   Map<RequestMethod, Map<String, RequestHandler>> _requestHandlers =
       Map.fromIterable(
     RequestMethod.values,
@@ -19,6 +26,7 @@ class DioAdapter extends HttpClientAdapter with MockProvider, Tracking {
     value: (element) => {},
   );
 
+  /// Takes in route, request, and sets corresponding [RequestHandler].
   RequestHandler onRoute(String route, {Request request = const Request()}) {
     _requestHandlers[request.method][route] = RequestHandler(route);
 
@@ -29,6 +37,7 @@ class DioAdapter extends HttpClientAdapter with MockProvider, Tracking {
     return requestHandler;
   }
 
+  /// [DioAdapter]`s custom configuration intended to work with mock data.
   @override
   Future<ResponseBody> fetch(RequestOptions options,
       Stream<List<int>> requestStream, Future cancelFuture) async {
@@ -41,9 +50,9 @@ class DioAdapter extends HttpClientAdapter with MockProvider, Tracking {
       if (exists) {
         history.advance();
 
-        return readMock(fullPath);
+        return _mockHandler.readMock(fullPath);
       } else {
-        return createMock(
+        return _mockHandler.createMock(
           fullPath,
           history.current.requestHandler
               .requestMap[history.current.requestHandler.statusCode],
@@ -52,6 +61,7 @@ class DioAdapter extends HttpClientAdapter with MockProvider, Tracking {
     });
   }
 
+  /// Closes the [DioAdapter] by force.
   @override
   void close({bool force = false}) => _adapter.close(force: force);
 }
