+++ lib/src/constants.dart
@@ -1,4 +1,7 @@
 import 'package:http_mock_adapter/src/utils.dart';
 
+/// [mockDirectoryPath] encompasses all kinds of mock data files in its path.
 const String mockDirectoryPath = 'assets/mocks';
+
+/// The root (working) directory of the project.
 String rootDirectoryPath = getRootDirectoryPath();
